# PERF: improve performance of infer_dtype